requirements.txt
matplotlib==3.8.2
numpy==1.26.2
opencv-python-headless
pandas==2.1.4
Pillow==10.1.0
streamlit==1.28.2